Finished implementation of pi3hat interface, time for tests

+++ src/meldog_hardware/pi3hat_hardware_interface/include/controllers/ControllerBridge.hpp
@@ -23,7 +23,7 @@ struct ControllerState
     double position_;   /* [radians] */
     double velocity_;   /* [radians/s] */
     double torque_;     /* [Nm] */
-    int temperature_;   /* [Celcius] */
+    double temperature_;   /* [Celcius] */
     bool fault = false;
 };
